refactor: move team hot encoding to its own folder
as more hot encoding is likely to take place it is
necessary to move the code to individual folders.
