+++ src/wrangling/encode/teams/hot_encode_teams.py
